+++ nevergrad/optimization/oneshot.py
@@ -3,6 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import copy
 import numpy as np
 from scipy import stats
 from scipy.spatial import ConvexHull  # pylint: disable=no-name-in-module
@@ -25,17 +26,27 @@ def convex_limit(struct_points: np.ndarray) -> int:
         return len(struct_points) // 2
     for i in range(0, min(2 * d + 2, len(struct_points)), 2):
         points += [struct_points[i]]
-    hull = ConvexHull(points, incremental=True)
-    k = len(points)
-    for i in range(d + 1, len(points)):
+    hull = ConvexHull(points[: (d + 1)], incremental=True)
+    num_points = len(hull.vertices)
+    k = len(points) - 1
+    for i in range(num_points, len(points)):
+        # We add the ith point.
         hull.add_points(points[i : (i + 1)])
-        if i not in hull.vertices:
-            k = i - 1
-            break
+        num_points += 1
+        if len(hull.vertices) != num_points:
+            return num_points - 1
+        for j in range(i + 1, len(points)):
+            # We check that the jth point is not in the convex hull;
+            # this is ok if the jth point, added in the hull, becomes a vertex.
+            hull_copy = copy.deepcopy(hull)
+            hull_copy.add_points(points[j : j + 1])
+            if len(hull_copy.vertices) != num_points + 1:
+                return num_points - 1
     return k
 
 
 def hull_center(points: np.ndarray, k: int) -> np.ndarray:
+    """Center of the cuboid enclosing the hull."""
     hull = ConvexHull(points[:k])
     maxi = np.asarray(hull.vertices[0])
     mini = np.asarray(hull.vertices[0])
@@ -147,8 +158,10 @@ class _RandomSearch(OneShotOptimizer):
         scale = self.scale
         if isinstance(scale, str) and scale == "auto":
             # Some variants use a rescaling depending on the budget and the dimension (1st version).
+            assert self.budget is not None
             scale = (1 + np.log(self.budget)) / (4 * np.log(self.dimension))
         if isinstance(scale, str) and scale == "autotune":
+            assert self.budget is not None
             scale = np.sqrt(np.log(self.budget) / self.dimension)
         if isinstance(scale, str) and scale == "random":
             scale = np.exp(self._rng.normal(0.0, 1.0) - 2.0) / np.sqrt(self.dimension)
@@ -300,8 +313,10 @@ class _SamplingSearch(OneShotOptimizer):
         if self._rescaler is not None:
             sample = self._rescaler.apply(sample)
         if self.autorescale is True or self.autorescale == "auto":
+            assert self.budget is not None
             self.scale = (1 + np.log(self.budget)) / (4 * np.log(self.dimension))
         if self.autorescale == "autotune":
+            assert self.budget is not None
             self.scale = np.sqrt(np.log(self.budget) / self.dimension)
 
         def transf(x: np.ndarray) -> np.ndarray:
@@ -383,13 +398,20 @@ MetaRecentering = SamplingSearch(
 MetaTuneRecentering = SamplingSearch(
     cauchy=False, autorescale="autotune", sampler="Hammersley", scrambled=True
 ).set_name("MetaTuneRecentering", register=True)
-HAvgMetaRecentering = SamplingSearch(
+HullAvgMetaTuneRecentering = SamplingSearch(
+    cauchy=False,
+    autorescale="autotune",
+    sampler="Hammersley",
+    scrambled=True,
+    recommendation_rule="average_of_hull_best",
+).set_name("HullAvgMetaTuneRecentering", register=True)
+HullAvgMetaRecentering = SamplingSearch(
     cauchy=False,
     autorescale=True,
     sampler="Hammersley",
     scrambled=True,
     recommendation_rule="average_of_hull_best",
-).set_name("HAvgMetaRecentering", register=True)
+).set_name("HullAvgMetaRecentering", register=True)
 AvgMetaRecenteringNoHull = SamplingSearch(
     cauchy=False,
     autorescale=True,