+++ nevergrad/parametrization/_datalayers.py
@@ -122,7 +122,7 @@ class BoundLayer(Operation):
 
     def set_normalized_value(self, value: np.ndarray) -> None:
         """Sets a value normalized between 0 and 1"""
-        bounds = tuple(b * np.ones(value.shape) for b in self.bounds)
+        bounds = tuple(b * np.ones(value.shape) for b in self.bounds)  # type: ignore
         new_val = bounds[0] + (bounds[1] - bounds[0]) * value
         self._layers[self._layer_index]._layered_set_value(new_val)
 
@@ -138,7 +138,7 @@ class Modulo(BoundLayer):
 
     def __init__(self, module: tp.Any) -> None:
         super().__init__(lower=0, upper=module)
-        if not isinstance(module, (np.ndarray, np.float, np.int, float, int)):
+        if not isinstance(module, (np.ndarray, np.float_, np.int_, float, int)):
             raise TypeError(f"Unsupported type {type(module)} for module")
         self._module = module